[TASK] Add LogRecordMessageTest
